Fix bug with one-word projects
Fix the bug that occurs when this template is used to create a new repository whose name is one word with no hyphens, such as oneword. In these instances, pytest fails to identify the submodules present and fails to import modules needed to test. Fix this by prefixing the pytest call with python -m. Closes #24