Added module names to pylint exceptions config
Updated the pylint control file after receiving this error in GitHub Actions:
"pylint: Command line or configuration file:1: UserWarning: Specifying exception names in the overgeneral-exceptions option without module name is deprecated and support for it will be removed in pylint 3.0. Use fully qualified name (maybe 'builtins.BaseException' ?) instead"

The fix is to prefix `builtins.` to the existing entries for this config option.